move AxisType in to core
